make dotdict a lib
