+++ LDAP/ldap2s3/dotdict/__init__.py
@@ -1,0 +1,1 @@
+from dotdict.dotdict import DotDict
